refactor(preact): move and update hooks
