--- hooks/__mocks__/settings.js
@@ -1,10 +1,0 @@
-export function useQueryProps() {
-  return {
-    isFullYear: false,
-    year: 2021,
-    month: 1,
-    shiftModel: "6-6",
-    group: 0,
-    search: null,
-  };
-}
