--- hooks/__mocks__/time.js
@@ -1,7 +1,0 @@
-export function useToday() {
-  return [2021, 1, 25];
-}
-
-export function useTodayZeroIndex() {
-  return [2021, 0, 25];
-}
