+++ lib/utils.js
@@ -97,7 +97,7 @@ export function getTodayUrl({ shiftModel, today, group }) {
 
 /**
  * Get today as an array of [yyyy, m, d, h] (all numbers).
- * @returns {[number,number,number,number]} [Year, month (start at 1), day, hour]
+ * @returns {[number,number,number,number,number]} [Year, month (start at 1), day, hour, minutes]
  */
 export function getToday() {
   const utcNow = Date.now();
@@ -133,4 +133,17 @@ export function parseNumber(str, fallback) {
   const num = parseInt(str, 10);
   const result = Number.isNaN(num) ? fallback : num;
   return result;
+}
+
+/**
+ * Handles click Events on elements with a title and displays them.
+ * @param {MouseEvent} event  MouseEvent of an click.
+ */
+export function titleAlertHandler(event) {
+  if (event.target instanceof HTMLElement) {
+    const title = event.target.closest("[title]")?.title?.trim() ?? "";
+    if (title) {
+      window.alert(title);
+    }
+  }
 }
